Backend/requirements.txt
appdirs==1.4.4
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.6.15
click==8.1.3
coverage==6.4.1
distlib==0.3.4
Django==4.0.5
filelock==3.7.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
mysqlclient==2.1.0
numpy==1.22.4
pandas==1.4.2
pep8==1.7.1
pipenv==2022.6.7
pycodestyle==2.8.0
pycparser==2.21
PyMySQL==1.0.2
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.37
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
virtualenv==20.14.1
virtualenv-clone==0.5.7
Werkzeug==2.1.2
